+++ dflib/src/test/java/org/dflib/ColumnSet_Select_ColsTailTest.java
@@ -1,0 +1,56 @@
+package org.dflib;
+
+import org.dflib.unit.DataFrameAsserts;
+import org.junit.jupiter.api.Test;
+
+public class ColumnSet_Select_ColsTailTest {
+
+    @Test
+    public void withinBounds() {
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                        1, "x",
+                        2, "y")
+                .colsTail(1).select();
+
+        new DataFrameAsserts(df, "b")
+                .expectHeight(2)
+                .expectRow(0, "x")
+                .expectRow(1, "y");
+    }
+
+    @Test
+    public void zero() {
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                        1, "x",
+                        2, "y")
+                .colsTail(0).select();
+
+        new DataFrameAsserts(df).expectHeight(0);
+    }
+
+    @Test
+    public void outOfBounds() {
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                        1, "x",
+                        2, "y")
+                .colsTail(4).select();
+
+        new DataFrameAsserts(df, "a", "b")
+                .expectHeight(2)
+                .expectRow(0, 1, "x")
+                .expectRow(1, 2, "y");
+    }
+
+    @Test
+    public void negative() {
+        DataFrame df = DataFrame.foldByRow("a", "b").of(
+                        1, "x",
+                        2, "y")
+                .colsTail(-1).select();
+
+        new DataFrameAsserts(df, "a")
+                .expectHeight(2)
+                .expectRow(0, 1)
+                .expectRow(1, 2);
+    }
+}